Load Config File
Loading the default configuration file for Electrs make possible to
authenticate in remote nodes using 'auth' directive, which can't be
passed as an option direct to the daemon, it's only accepted in the
'config.toml' file.

This file must be in the HOST_DATA_DIR to get read by Electrs as it
can be used to pass extra configurations that are not in the '.env'.